fix flakiness in FastForwardHttpReporterTest
the current test asserts that, after the reporter is started,
`httpClient.sendBatch(..)` should be called at least twice. Since this
method is called via something scheduled with a
ScheduledExecutorService, the test can't 100% be sure how many times it
will be called within a given time window. Sometimes this test fails
because the method was called only once within the timeout, not at least
twice.

This change changes the assertion to be `atLeastOnce`, since the logic
being tested does not really depend or care about how many times
`httpClient.sendBatch(..)` is called - testing ScheduledExecutorService
is not the goal of this test.