adjust the warning about MetricId being expensive / MetricIdCache
Some number of people who read this end up reaching for the extra complexity of a MetricIdCache before even knowing that their application is spending considerable time simply constructing/copying MetricId instances. This change softens the warning and removes the assertion that "MetricId is expensive", which doesn't offer any quantification.