Make count decrease align with increase.
The `State.update` always increase `count` regardless if it is at full `size` or not. Which means it can grow larger than `size`. So when `backfill` replaces the `values` with an empty `ConcurrentSkipListMap` and only decrement by `-oldValues.size()`, it means count can still be larger than size while map is empty. Which can cause a NoSuchElementException at next update.