Don't report empty histograms to FastForward (#140)
*Rationale*

If histograms don't have any reported values, or have not been used
in a long time, the snapshot is empty and the histogram metrics
will not have a reasonable default value. All stats (min, max, median, etc) would
incorrectly be reported as zero

*Performance impact*

The performance impact should be minimal, since it only adds one additional
if-check when reporting histograms. Snapshot.size() is a fast method, simply
returning the length of its internal array.

For cases where snapshots are empty - i.e. unused histograms we should see
a performance improvement since we can skip emitting the data points to FastForward
as well as skipping the creation of the metric ids.